can i write  while loop

+++ bin/trim_circular.py
@@ -14,8 +14,10 @@ def main():
   cc_fn = sys.argv[1]
   res = trim_circular(cc_fn)
   next_res = res
-  while next_res != '':
+  while True:
     next_res = trim_circular(res)
+    if next_res == '':
+      break
     status = commands.getstatusoutput('rm -rf ' + res)[1]
     status = commands.getstatusoutput('mv ' + next_res + ' ' + res)[1]
   return
